tile rendering

+++ src/underground_level.cpp
@@ -1,6 +1,6 @@
 #include "underground_level.hpp"
 
-MppTileType MppUndergroundLevel::Generate(int x, int y) const
+MppTileType MppUndergroundLevel::Generate(float x, float y) const
 {
     return MppTileTypeGrass;
 }
